Update patch set 1
Patch Set 1:

> Note: there are more users than ART. E.g., libdex, fastboot, ...,
 > use it as a static library and do not import libz-host, either. Do
 > they need to be changed, too?

Not sure, and don't know how to test (other than get a Mac, set it up, etc - sounds like a day of work). If a target only uses libziparchive, and not libz, then it's OK; otherwise we should add libz. Probably most or all of those don't need libz.

How about I land this tomorrow morning and keep an eye on the build server to quickly fix remaining issues? Or it would be great if someone tested a checkbuild on Mac.

Patch-set: 1