Update patch set 1
Patch Set 1:

I believe the power supply uevent needs to be a wakeup source in order to ensure that we switch to the faster poll time.  Without a wakeup source, plugging the device while it was in suspend could cause it to start charging, wake up to deliver the uevent, but return to suspend before handling the uevent and not increase the poll rate.

If you are unable to suspend due to uevents that suggests your kernel is sending too many uevents.  You generally want to minimize the uevents sent by the kernel anyways because every uevent wakes up every process that is listening for them (ueventd, healthd, framework, etc.)

Patch-set: 1