Update patch set 1
Patch Set 1:

The charger driver should not be holding a wakelock, we want to suspend while charging.

For wakeup events, the driver needs to hold a wakelock until it has passed the event to userspace, at which point userspace needs to hold the wakelock until it is done processing the event.  EPOLLWAKEUP handles holding the wakelock while userspace is processing the event.

If the uevent listener doesn't care about timely processing of the uevent, then it doesn't need EPOLLWAKEUP.  In this case, we need to make sure that we have set the wakealarm to the shorter poll interval before entering suspend.

Patch-set: 1