adb disconnect: fix write-after-free memory corruption and crash
Transport atransport objects are semi-reference counted: the input and
output threads each hold a reference. The adb disconnect command was
calling transport_unref to release a reference that it never had in the
first place. This meant that the refcount dropped to zero and the object
was deleted before either the input or output thread released its
reference. When that last thread released its reference, it wrote to
freed memory and also sometimes crashed.

The fix is to not release any unheld reference, to only deal with
transport objects when the proper locks are held, and to let the
"natural" transport removal process (in transport_registration_func)
remove the transport from transport_list.

Also in this change:

 - Add unregister_tcp_transports which replaces
   unregister_all_tcp_transports and find_transport +
   unregister_transport.

 - An alternative would be to fix find_transport to increment the
   refcount and to use it in adb disconnect, but the downside is that
   this would be a more risky change because then the refcount might
   drop to zero on the adb disconnect thread instead of the input or
   output thread, which is already the common case.

 - CHECK_GT the refcount before decrementing. This would have caught
   the bug in the first place.

 - Add some clarifying comments and BUGBUG notes for bugs that I found
   while reading the code.

 - Add some symbolic constants to make some code easier to read.

Change-Id: I216edbfc5c677f5c58a6f138532ffec675296710
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>