Update patch set 1
Patch Set 1:

> ah. maybe we should use German instead then. this was the other
 > thing i was going to say: "adb shell ls -l" for a device that
 > already had some korean filenames just showed "???".

Yeah, it is a real pain because people might file bugs on you when the answer is that they need to reconfig their machine (as far as I can tell). It would be nice to at least have adb give a warning in this scenario, but it seems messy -- how would that even work? We'd have to parse all output, look for CJK, check the Windows registry for system settings, then where could we output a warning to? Mid-stream in output? That sounds awful. Maybe just document this on developer.android.com and eventually the word will get out?

 > so it's just CJK that doesn't work? Windows is weird.

Yeah, I think 'East Asian' languages need special fonts in the console. I'm a Windows person and I've got to say that the Windows console feels very, very old, almost unmaintained (though MS has made some improvements for Windows 10, but none for this Unicode business).

 > i'll have a look on monday.

Yes, time for a break. :)

Patch-set: 1