[WIP] Faster adb sync.
This quick hack shows that we can double the speed of "adb sync" (on N9).
If we remove the constant limit and use the per-transport limit introduced
recently to speed up pull/push, this is another improvement we can ship
without protocol changes (though again, it will only help devices with a
newer adbd).

Change-Id: If9f6d4c1f93ec752b95f71211bbbb1c513045166