Update patch set 10
Patch Set 10: Verified+1

This is ready for review now, except for removing the init_parser_test, which I have a question about...  

Is there any easy way to run the tests that check for seclabel_, uid_, gid_, etc, where I don't have or need getters for those internal variables?  In theory I could create a friend out of the test classes, but that seems ugly.  It seems to me just as ugly to have to create a long list of getters strictly for testing.

This, combined with the change for Action/Command classes allow us to remove the infamous strdup() in parse_config(), though I want to do that in a separate change.

Patch-set: 10
Label: Verified=+1