liblog: fix win32 build
c744ef55478d7f4da37cb040f901cf3e9ec75dbb added LOCAL_CXX_STL=none, but
that broke the win32 build because it uses the linker flag
-nodefaultlibs. That prevents the Windows C Runtime (MSVCRT.DLL) from
being linked to, but that is needed by liblog on Windows because it
calls read(), write(), etc. which are implemented by MSVCRT.DLL on
Windows.

Change-Id: I6b322b4eebd4f1473da1308b7e984e349d64fc4b
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>