Update patch set 11
Patch Set 11:

(23 comments)

I've been keeping the copyright the same as the other files including the /* */ as well as the opening brace on the new line, though I guess we can break from this as we bring this up to the C++ coding guidelines?

Should we start renaming function from call_this_func() to CallThisFunc() as we touch them as well?  We'll temporarily have mixed style in the legacy code, but that may be better?  I'm thinking about this for functions such as msg_start() that I modify though am not restyling yet.

Patch-set: 11
Label: Verified=0