Update patch set 2
Patch Set 2:

Yes, the current patch is no good. I agree with the proposal to have a separate fd for this.

About the only part that bothers me is that I don't want this command line option to be visible - it shouldn't be used by anything other than this internal comunication logic. If there are already options that do that, then let me know and I'll follow that pattern, otherwise I guess it would have to be a user visible option..

Patch-set: 2
Label: Verified=0