Update patch set 1
Patch Set 1:

This is a bit out of my expertise, I think Dan is the right person to review this.

My hunch was that the reason the HOST global existed was because the same code is compiled into libadb and/or libadbd, which might be linked into a HOST or non-HOST binary. So that's why it can't be done with an #ifdef. But this was always just my assumption, I never fully analyzed the code to verify if this is really the case.

I also found it confusing the setting of HOST in init_usb_transport and the complicated looking code in local_init.

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0