throttle automatic service respawning
This greatly increases the time required to defeat statistical defenses
like ASLR via bruteforce in auto-restarted services like mediaserver. It
calculates the time since the service was started in the signal handler
and then enforces the rate limit by sleeping in the forked child before
executing the service.

This was inspired by grsecurity's GRKERNSEC_BRUTE feature.

Change-Id: I5166915a84a7b47ea37fbe974aa9191417577555