adb: remove race between main thread and device_poll_thread.
If running `adb kill-server`, the main thread will call exit(),
which calls global variable's destructors. But these global variables
may be using by other threads. One example is g_usb_handles used
in usb_linux.cpp, normally it is protected by lock, but it is not
protected by lock when the destructor is running. Change exit()
to _exit() may be the easiest way to fix this, but it may cause
other problems.

Bug: 23384853
Change-Id: I3119d1d4ef38aef6a86ebba22f4bed2e0264a85e