Update patch set 6
Patch Set 6:

This is ready for review.  It's tested on N9 and I'm testing other targets now.

There is admittedly a lot going on in this change.  I thought it was better to continue with removing the old Keyword struct in this single change than creating an interface into lookup_keyword(), etc as is done in patch set 2.  Let me know if it's better for everyone that I break it up.

Patch-set: 6
Label: Verified=0