adb: remove kCsNoPerm.
kCsNoPerm is a special usb state showed only on linux host. It said
to happen when there is no permission to write to usb device. I prefer
to remove this state because:
1. I have never met with this state. I think it is rare or doesn't exist
anymore.
2. It is not supported on mac and windows. And the only function of it
is to notify the user that something is wrong with your usb connection
when running `adb devices`. So I think it is not helpful to users.
For developers, they can find details in the log.
3. It keeps the code path of having a non readable and non writable
transport in the transport list, which has no input_thread or output_thread.
Actually, this types of transport have no way to release.

After all, I think kCsNoPerm state doesn't deserve being kept anymore.

Change-Id: Ib76cfabd0b7d4c7fbb1ffc4311f31c83d86429f6