Update patch set 2
Patch Set 2:

Unfortunately, I can't easily add a test for this, so I wound up modifying the code to force the signal handler not to do anything to test it. I then also forced a secondary signal to be sent to verify I can see the spurious signal message.

Patch-set: 2