Update patch set 1
Patch Set 1:

In the emulator code, adb-qemud.c relies on hw-qemud.c. When there is a host disconnect,
hw-qemud.c does't close the connect with the guest, instead it sends "disconnect:00" message.
adb-qemud.c overrides this behavior and sends A_SYNC(0,0) instead. Besides the method used
in this patch, we can change the behavior of hw-qemud.c to just close the connect with the
guest. The latter is simpler and should be better. But that means a lot of testing work.
I think I should find someone more familiar to android emulator.

Patch-set: 1