Update patch set 2
Patch Set 2:

I agree, but logcat can not tolerate nulls in embedded string content either ... we would probably be advised to rewrite liblog/logprint.c in C++.

This problem has occurred on corruption of sys/fs/pstore/console-ramoops (outside this venue) and on Mediatek kernels. klogd is not released in "user". Such hardening would be required if we decided to make klogd mainstream.

Patch-set: 2