Update patch set 1
Patch Set 1:

i've been promising to port this from art for a long time. this adds the extension (that's been useful in toybox) of letting the caller constrain min and max values smaller than the type. no support for hex, but i'm assuming we'll add ParseHexUint when we need it (and tbao doesn't need it in the recovery image iirc).

Patch-set: 1