Update patch set 2
Patch Set 2:

devices.cpp:device_init() is walking on the sysfs uevent files to
force the regeneration of the uevent so ueventd will create the dev
nodes for the devices that it could have missed before it was started.
It does not guarantee, that once we get out of device_init() all the
device nodes are created.

Here is the comment on do_coldboot()
/* Coldboot walks parts of the /sys tree and pokes the uevent files
** to cause the kernel to regenerate device add events that happened
** before init's device manager was started
**
** We drain any pending events from the netlink socket every time
** we poke another uevent file to make sure we don't overrun the
** socket's buffer.
*/

devices.ccp:device() creates the .coldboot_done file once it has
walked on the sysfs and not once all the uevent have handled.
Meaning, when .coldboot_done file is created, the uevent are not yet
handled and the dev nodes are not necessarily created.

In init.c:main():
[...]
// Queue an action that waits for coldboot done so we know
ueventd has set up all of /dev...
[...]

I think this comment is wrong.  .coldboot_done ensures that we get
back all the uevents we might have missed not that all the devices
have been probed and dev node created.

BTW, if .coldboot_done meant what init.cpp is expecting, init.cpp
would not do :
    // ... so that we can start queuing up actions that require stuff from /dev.
    queue_builtin_action(mix_hwrng_into_linux_rng_action, "mix_hwrng_into_linux_rng");
    [...]
    // Repeat mix_hwrng_into_linux_rng in case /dev/hw_random or /dev/random
    // wasn't ready immediately after wait_for_coldboot_done
    queue_builtin_action(mix_hwrng_into_linux_rng_action, "mix_hwrng_into_linux_rng");

Patch-set: 2