fs_mgr: Fallback to empty suffix if A/B suffix cannot be determined.
Instead of falling back to suffix _a, we now fall back to the empty
suffix if neither the kernel commandline nor the misc partition
specifies the suffix.

This CL the advantage that existing devices will keep working without
repartitioning, avoiding (costly) flag day. The disadvantage is that the
repartitioning script will have to write out e.g. _a in the misc
partition.

Change-Id: I3f58928a664433504ebdf8d0ee05a319be5097cf