Update patch set 2
Patch Set 2:

> > I'm glad you've reduced the timeout from 5 seconds to 1 second in
 > > init.cpp otherwise we would have kept this bug for a very long
 > > time.
 > 
 > excellent! this isn't the first bug that's been found this way, and
 > i doubt it will be the last.
 > 
 > (i do wonder whether init should just commit suicide instead? maybe
 > that would make problems like this easier to debug?)

It would make bug like this a lot easier to catch and I'm generally in favor of an assert approach.  Especially for early code like this one that should not have to be fault tolerant.

In the meantime, I will work with our test team to get the init log "Timed out waiting for /dev/.coldboot_done" monitored to ensure that no patch will break it once the culpirit driver will be fixed.

Patch-set: 2