adb unittest: fix Windows Unicode
adb.shell() was recently changed to use subprocess.Popen(), which
doesn't work properly with Unicode on Windows. The fix is to use the
same work-around that I did for subprocess.check_output(): write UTF-8
to a batch file and run it. The change is primarily refactoring to
enable code reuse.

Change-Id: Icac8bacbb5e93cbcc587def58a5f55b03bb6239f
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>