adb: add support for feature versions.
Adds standardized ways to specify and check supported feature versions.
This will make it easier to evolve features without breaking backwards
compatibility.

The next CL will use this to bump the shell version to 3 in order to
add support for manually selecting raw or PTY shells.

One side effect of this CL is that it will temporarily disable the
shell protocol with existing devices until they are rebuilt. This is
done now to avoid breaking the shell protocol logic once we bump
the shell version to 3. Since no devices have been release with the
existing shell protocol I don't think this will be a problem.

Change-Id: Ice18cd292063d02720ec2ff6c364401080601316