Update patch set 3
Patch Set 3:

Either way seems OK to me, since adbd created the screencap process it makes sense that it should kill it when it's no longer needed (as long as adbd has the selinux permissions to send SIGKILL).

In general restoring signal handlers after forking seems like it might be a good idea as well, but as far as I know at the moment we don't do this for `adb shell` either.

Patch-set: 3