fastboot: add 'provision' sub-command.
Without a file argument, e.g.

 $ fastboot provision

we'll look for ${ANDROID_PRODUCT_OUT}/provisioning-commands (with a file
argument we'll use the given file). The file is simply a list of
fastboot commands to send to the device, for example

 flash partition ${ANDROID_BUILD_TOP}/device/.../partition-tables.bin
 flash vendorXYZ ${ANDROID_BUILD_TOP}/device/.../vendorXYZ.bin
 ...
 flash boot_a ${ANDROID_PRODUCT_OUT}/boot.img
 flash system_a ${ANDROID_PRODUCT_OUT}/system.img
 flash userdata ${ANDROID_PRODUCT_OUT}/userdata.img

The environment variables ANDROID_BUILD_TOP and ANDROID_PRODUCT_OUT can
be used in the provisioning-commands file.

This mechanism can be used to replace all partitions with known
versions (e.g. provision the device). The idea is that board-specific
build scripts can generate the file including customizing it for
different build configurations (for example, whether A/B is used or
not).

Change-Id: Iffe36cb26ea41321a5fc82db5e5c4c3524975754