libcutils: reset PR_SET_TIMERSLACK_PID value to avoid conflict
PR_SET_TIMERSLACK_PID value keep colliding with that of newer
prctls in mainline kernel and/or in kernel headers in bionic
(e.g. first with PR_SET_THP_DISABLE, and again with
PR_MPX_ENABLE_MANAGEMENT).

So reset PR_SET_TIMERSLACK_PID to a large number so as to
avoid conflict in the near term while it is out of mainline
tree.

Corresponding Change-Id up for review in kernel/common is
I061b25473acc020c13ee22ecfb32336bc358e76a

Change-Id: Icd8c658c8eb62136dc26c2c4c94f7782e9827cdb
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>