DO NOT SUBMIT system/core: poc for copy-on-write data structures
The copy constructors, operator =, etc of String8, String16 and Vector
are used heavily. Eliminating SharedBuffer and making those types wrappers
of std::string and std::vector might result in unpleasant surprises
wrt memory usage.

I think I found a way to still have the "copy-on-write" semantics
provided by SharedBuffer just with wrappers and with minimal effort,
by having a class that provides copy-on-write semantics for any class,
and then having wrappers of copy-on-write objects.

Please let me know if you think this approach is viable (this is just
a poc, haven't tried to compile even).

Change-Id: Ia81ae71b6f993e4d6dc0e8553f04e7a84cb27d43