Update patch set 1
Patch Set 1:

There is a non-insignificant amount of code in the Android tree that does this exact fragment. Android init does this kind of thing too with rampant abandon. I will accept the lack of comments as a code standard with regards to this. The _first_ speedup that showed up in the SysVr3 kernel was strcmp fragments like this (they actually replaced strcmp and its ilk with a STRCMP macro though). I am expecting this is common knowledge, and in this case I will defer to git blame to point the errant person as to why. log_is_loggable.c is probably the _most_ optimized piece of code in the Android ecosystem (thumping his chest) ;-}

Patch-set: 1