Update patch set 2
Patch Set 2:

in answer to a question from ccross... increasing the packet size (i've already fixed adbd to cope with arbitrarily large packets as long as we can malloc space for them) from 64KiB to 8MiB only shaves about 1s off the total time. (despite cutting the number of reads and writes in the client drastically.) my _assumption_ is that we're already fast enough at this point that we're limited by how fast we can write on the device side.

Patch-set: 2