Update patch set 2
Patch Set 2:

> crash_reporter currently uses metrics_lib.Init().  Should I also be
 > changing this to InitNoCache()?

For the client I think it is fine to use Init. It will give you a wrong answer only if you call twice in the same second and the state changed in between. I don't think we care about having a granularity that fine.

This amount of ("expensive") precision is only needed in metrics_daemon.

Patch-set: 2