Update patch set 1
Patch Set 1:

I added a comment on how to use it. But who could possibly use it as in your second suggestion? If there are no elements it.value() would be undefined, and if you checked before the "do {" using next(), you would be losing the first one in case there were elements...

Patch-set: 1
Label: Verified=0