The interface for adb backup was changed, but that breaks backup for all android devices in the world.
The backup manager service expects adb to send backup commands
in the format:

backup:<arg>*

In other words colon separated arguments with no spaces, or quotes.

However adb was changed to remove the colons, add spaces and quotes.
This breaks adb backup for all shipped android devices, and produces
a backup file that is 41 bytes and contains no actual backup data.

The only real indication that something went wrong with the backup
is a logcat message like this:

W/BackupManagerService(  798): Unknown package  '-system' 'android' 'com.android.providers.settings', skipping

To test:
$ adb devices
List of devices attached
0421c368d694f521	device
$ adb backup -all
Now unlock your device and confirm the backup operation.
$ ls -l backup.ab
-rw-r----- 1 desjardi desjardi 41 Oct 15 09:44 backup.ab

Note that the backup.ab is 41 bytes, normally it is kilobytes.
Restoring this backup.ab will appear to work, except that nothing
will be restored on the device.

Change-Id: Ie90c774aff81da449da005b1d7bdb68f5da7fab4