Adb: fix adb backup compatibility.
Originally adb backup command used colon to separate arguments, and it
sent command string like "backup:-apk:-all:-system" to device.
But now it changes to use space and apostrophes to separate arguments, and
it sends command string like "backup '-apk' '-all' '-system'" to device.
So new adb can't be used for old devices, and old adb can't be used for new devices.
To make it compatible with old adb and old devices, we need to change back to
use colon as argument separator.

Bug: 24976399

Change-Id: Ie90c774aff81da449da005b1d7bdb68f5da7fab4