healthd: BatteryMonitor should monitor type unknown
The BatteryMonitor should monitor power sources that at boot are unknown.
The problem that may happen otherwise is that if the device is connected
to some unknown USB charger at boot it will not acknowledge proper
chargers later.
From the users pov this means that the charging flash indication in the
battery icon never appears.

Change-Id: Ieb3c31223206320173f5b15c0093d21287965fe2