Update patch set 1
Patch Set 1:

this is the version ccross doesn't like because it doesn't show _overall_ percentage. but it's already a big enough change to be a good review point.

(i think doing overall percentage is going to require changing SyncConnection to have more intention-revealing methods than just "Print", so it can construct the full status line from the various things it knows.)

Patch-set: 1