fastboot: create Transport object.
This CL creates a Transport object to provide a generic interface for
various transports. Specifically this is designed to be able to add UDP
support to fastboot without changing the program logic.

In order to stay reasonably sized, this CL just moves existing
functionality to the Transport interface, using existing names and code
structure when possible.

Bug: http://b/22029765
Change-Id: I68641af0da7d13db4647f5e20a18d04d67f0b327