Update patch set 2
Patch Set 2:

another day, another way to make spencer sad...

this works great on Unix, but the internets lead me to believe there's no such thing as EINTR on Windows, so i guess the solution there is going to have to look very different. in which case i wonder if there's one implementation of the "interrupt read when the window size changes" that will work for both?

(ps2 adds some more comments.)

Patch-set: 2