Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > > another day, another way to make spencer sad...
 > >
 > > this works great on Unix, but the internets lead me to believe
 > > there's no such thing as EINTR on Windows, so i guess the
 > solution
 > > there is going to have to look very different. in which case i
 > > wonder if there's one implementation of the "interrupt read when
 > > the window size changes" that will work for both?
 > 
 > I think we can get this to work for Windows. Basically, when
 > unix_read() calls ReadConsoleInput(), it will get a window resize
 > event, and then we can make unix_read() return EINTR (which I think
 > is a constant even on Windows). Then edit the #ifdef _WIN32 part of
 > send_window_size_change() to call GetConsoleScreenBufferInfo() to
 > get the console size and send that along.
 > 
 > I haven't tried this, but I think it will probably work.

awesome. i was expecting you to say we were going to need to switch to select instead of read and having a SIGWINCH pipe or something.

 > What really makes spencer sad is when adb push shows a weird box
 > character on Windows :-/ :)

yeah, that's next-ish on my list, honest. it's just process politics --- get all your features in, even if they're a bit broken, so that fixing them becomes "just" a bug rather than a feature, and so less likely to be bumped if it does slip. doing this either makes me a bad person (if you're, say, a release manager) or a good person (if you're, say, an end user). today i'm wearing my gray hat, like always.

Patch-set: 2
Label: Verified=0