adb: win32: fix adb shell
28416d6d33267f92768116e269966d61eac01db9 broke adb shell on Windows
because it was calling unix_read() with a buffer size of 1 byte which
isn't supported by unix_read()/_console_read().

This change switches back to bigger reads by removing and inserting
characters from the bigger buffer when processing it.

Also modify _console_read() to CHECK() if the caller passes a very small
buffer size to detect this sort of issue faster in the future.

Change-Id: I8f0aed4fb9f6f2f5b9a6b68ce60d2e368fec81c7
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>