Add support for device class rules in ueventd.*.rc
ueventd does not support rules referring to symlinks, as class
attributes are.
For Ara it is necessary to be able to specify permissions using
device classes, to abstract away from platform details.

This patch adds support for rules (in ueventd.*.rc) in the form of
/sys/class/SUBSYSTEM/WILDCARDS ATTRIBUTE UMASK USER GROUP

wildcards are currently supported only at the end of the rule;
wildcards interleaved with path components (as
/sys/class/leds/*/brightness) are not supported.

Signed-off-by: Jacopo Mondi <jacopo.mondi@linaro.org>
Signed-off-by: Michael Scott <michael.scott@linaro.org>
Change-Id: I75b5d6128de8b90f13cbd35d23d2357ba1141e33
Signed-off-by: Rob Herring <robh@kernel.org>
Signed-off-by: Jacopo Mondi <jacopo.mondi@linaro.org>