Update patch set 2
Patch Set 2:

> > Looks like we don't actually have quick_exit on Windows. I'm
 > going
 > > to make all of our globals non-destructing instead.
 > 
 > Will "#define quick_exit _exit" (or equivalent) work ok enough on
 > Windows? I don't see any atexit handlers that actually need to run
 > on Windows?

Yeah, it should, but there's a comment that implies that we probably want to use an atexit handler on Windows at some point ("// TODO: can we use atexit? even on Windows?")

It looks like we only have 6 instances of globals with nontrivial destructors, and none of them need to be destructed, so I think that's a better approach.

Patch-set: 2
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0