Update patch set 1
Patch Set 1:

+1 for going into libbase.

libchrome does scoped files by defining a functor for their scoped_ptr type (which is like a unique_ptr before it became part of the standard.  We can't grow dependencies on libchrome elsewhere in Android, and a dependency on scoped_ptr is especially bad.

The implementation in libnativehelper is primitive enough that maybe we should update that file with a comment pointing to the libbase version.

This header could use a lot of polish.

Patch-set: 1
Label: Code-Review=0