Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

Will respin the API for __android_is_loggable in a compatible manner that hopefully will look less like a hack. Will drop bool return for lock functions, as we will decide not to support errorcheck type pthread_mutex, and defer support for it to the future.

Patch-set: 3
Label: Code-Review=-1