Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

I was fixing three problems with this code, one is the API abuse, another was the notification to understand that calling these functions carries a risk in a signal handler, and finally a fix for the deadlock we saw.

I did not fix the first problem because the mask is not enough, I need to regroup and decide if I want a KISS change for the deadlock condition, and address the other two as technical debt, or fix them all correctly.

Patch-set: 2
Label: Code-Review=-1