Fix adb sync error reporting.
We're now able to send packets faster than the device can handle them,
meaning that sometimes we're several packets through before the device
says "hey, wait, I can't write" and closes the connection. At best this
led to us reporting that we couldn't sync because "Connection reset";
at worst we'd get SIGPIPE because we were still streaming to a connection
that had already been closed.

I think the best solution is to always try to read the server's response
at the "end" of the transfer (even if that happens early because a write
fails because we get ECONNRESET).

Also rename adb_main adb_server_main, and move the ignoring of SIGPIPE
into adb_commandline so it applies to both client and server (but not
adbd).

Bug: http://b/25230872
Change-Id: Ic60e4d13ed03fdcdf0d5cbc97201ebd1097c16ed